fix code and document
