+++ deepmd/entrypoints/__init__.py
@@ -1,3 +1,4 @@
+# SPDX-License-Identifier: LGPL-3.0-or-later
 """Submodule that contains all the DeePMD-Kit entry point scripts."""
 
 from ..infer.model_devi import (
@@ -14,6 +15,9 @@ from .doc import (
 )
 from .freeze import (
     freeze,
+)
+from .gui import (
+    start_dpgui,
 )
 from .neighbor_stat import (
     neighbor_stat,
@@ -40,4 +44,5 @@ __all__ = [
     "make_model_devi",
     "convert",
     "neighbor_stat",
+    "start_dpgui",
 ]
