+++ deepmd_cli/__init__.py
@@ -1,0 +1,6 @@
+# SPDX-License-Identifier: LGPL-3.0-or-later
+"""This module contains the entry points for DeePMD-kit.
+
+If only printing the help message, this module does not call
+the main DeePMD-kit module to avoid the slow import of TensorFlow.
+"""
